+++ src/consumer/component/add-components/add-components.js
@@ -82,9 +82,10 @@ export type AddProps = {
   id?: string,
   main?: PathOsBased,
   namespace?: string,
-  tests: PathOrDSL[],
-  exclude: PathOrDSL[],
-  override: boolean
+  tests?: PathOrDSL[],
+  exclude?: PathOrDSL[],
+  override: boolean,
+  writeToBitMap?: boolean
 };
 
 export default class AddComponents {
@@ -96,7 +97,8 @@ export default class AddComponents {
   namespace: ?string;
   tests: PathOrDSL[];
   exclude: PathOrDSL[];
-  override: boolean;
+  override: boolean; // (default = false) replace the files array or only add files.
+  writeToBitMap: boolean; // (default = true)
   warnings: Object;
   ignoreList: string[];
   gitIgnore: any;
@@ -107,9 +109,10 @@ export default class AddComponents {
     this.id = addProps.id;
     this.main = addProps.main;
     this.namespace = addProps.namespace;
-    this.tests = addProps.tests;
-    this.exclude = addProps.exclude;
+    this.tests = addProps.tests || [];
+    this.exclude = addProps.exclude || [];
     this.override = addProps.override;
+    this.writeToBitMap = addProps.writeToBitMap !== undefined ? addProps.writeToBitMap : true;
     this.warnings = {};
   }
 
@@ -294,10 +297,6 @@ export default class AddComponents {
     const componentsWithFilesP = Object.keys(componentPathsStats).map(async (componentPath) => {
       if (componentPathsStats[componentPath].isDir) {
         const relativeComponentPath = this.consumer.getPathRelativeToConsumer(componentPath);
-        const absoluteComponentPath = path.resolve(componentPath);
-        const splitPath = absoluteComponentPath.split(path.sep);
-        const lastDir = splitPath[splitPath.length - 1];
-        const nameSpaceOrDir = this.namespace || splitPath[splitPath.length - 2];
 
         const matches = await glob(path.join(relativeComponentPath, '**'), {
           cwd: this.consumer.getPath(),
@@ -317,6 +316,10 @@ export default class AddComponents {
         const resolvedMainFile = this._addMainFileToFiles(files);
 
         if (!finalBitId) {
+          const absoluteComponentPath = path.resolve(componentPath);
+          const splitPath = absoluteComponentPath.split(path.sep);
+          const lastDir = splitPath[splitPath.length - 1];
+          const nameSpaceOrDir = this.namespace || splitPath[splitPath.length - 2];
           const idFromPath = BitId.getValidBitId(nameSpaceOrDir, lastDir);
           finalBitId = this._getIdAccordingToExistingComponent(idFromPath.toString());
         }
@@ -450,7 +453,7 @@ export default class AddComponents {
         if (addedResult) addedComponents.push(addedResult);
       }
     }
-    await this.bitMap.write();
+    if (this.writeToBitMap) await this.bitMap.write();
     return { addedComponents, warnings: this.warnings };
   }
 }